+++ .gitignore
@@ -1,3 +1,30 @@
+*.bak
+*.o
+.*.swp
+*.orig
+.swp
+docs/
+output/
+ojam/
+out/
+16000000/
+8000000/
+out_native/
+version.h
+Session.vim
+*.so
+*.so.*
+*.dylib
+*.dylib.*
+.DS_Store
+Makefile.inc
+utility/includes.h
+examples_linux/*
+!examples_linux/**/
+!examples_linux/*.*
+examples_linux_c/*
+!examples_linux_c/**/
+!examples_linux_c/*.*
 /examples_xc8/pingpair_dyn.X/build/sdcc/
 /examples_xc8/pingpair_dyn.X/dist/sdcc/
 /examples_xc8/pingpair_dyn.X/build/xc8/
