+++ examples/pingpair_dyn_c/pingpair_dyn_c.ino
@@ -13,7 +13,7 @@
  */
 
 #include "nRF24L01.h"
-#include "RF24_c.h"
+#include "RF24.h"
 
 //
 // Hardware configuration
