+++ examples_linux/transfer.cpp
@@ -16,7 +16,7 @@ TMRh20 2014
 #include <iostream>
 #include <sstream>
 #include <string>
-#include <RF24_c/RF24_c.h>
+#include <RF24/RF24.h>
 #include <unistd.h>
 
 using namespace std;
@@ -36,7 +36,7 @@ using namespace std;
 //RF24 radio(RPI_BPLUS_GPIO_J8_15,RPI_BPLUS_GPIO_J8_24, BCM2835_SPI_SPEED_8MHZ);
 
 // Setup for GPIO 15 CE and CE0 CSN with SPI Speed @ 8Mhz
-RF24_c radio(RPI_V2_GPIO_P1_15, RPI_V2_GPIO_P1_24, BCM2835_SPI_SPEED_8MHZ);
+RF24 radio(RPI_V2_GPIO_P1_15, RPI_V2_GPIO_P1_24, BCM2835_SPI_SPEED_8MHZ);
 
 /*** RPi Alternate ***/
 //Note: Specify SPI BUS 0 or 1 instead of CS pin number.
