+++ examples_linux_c/gettingstarted.c
@@ -24,7 +24,7 @@ TMRh20 2014 - Updated to work with optimized RF24 Arduino library
 //#include <sstream>
 //#include <string>
 #include <unistd.h>
-#include <RF24_c/RF24_c.h>
+#include <RF24/RF24.h>
 
 //
 // Hardware configuration
