+++ examples_linux_c/gettingstarted_call_response.c
@@ -15,7 +15,7 @@ TMRh20 2014 - Updated to work with optimized RF24 Arduino library
 
 #include <string.h>
 #include <unistd.h>
-#include <RF24_c/RF24_c.h>
+#include <RF24/RF24.h>
 
 
 //
