+++ examples_xc8/pingpair_dyn.X/nbproject/Makefile-genesis.properties
@@ -1,11 +1,11 @@
 #
-#Wed Dec 07 20:23:19 BRST 2016
+#Fri Jan 27 17:06:48 BRST 2017
 sdcc.languagetoolchain.dir=/usr/bin
 xc8.com-microchip-mplab-nbide-toolchainXC8-XC8LanguageToolchain.md5=1a80c324d373be944aab8c3a16e71c0c
 xc8_P16F877A.com-microchip-mplab-nbide-toolchainXC8-XC8LanguageToolchain.md5=1a80c324d373be944aab8c3a16e71c0c
 conf.ids=xc8,sdcc,xc8_P16F877A
 xc8_P16F877A.languagetoolchain.version=1.38
-configurations-xml=8b75ac8a7b89a05ffa495118a63fe17f
+configurations-xml=7b933ab4f4263858d8203b43030a1bce
 sdcc.languagetoolchain.version=3.6.0
 xc8.languagetoolchain.dir=/opt/microchip/xc8/v1.38/bin
 xc8_P16F877A.languagetoolchain.dir=/opt/microchip/xc8/v1.38/bin
