+++ examples_xc8/pingpair_dyn.X/nbproject/Makefile-xc8.mk
@@ -51,17 +51,17 @@ OBJECTDIR=build/${CND_CONF}/${IMAGE_TYPE}
 DISTDIR=dist/${CND_CONF}/${IMAGE_TYPE}
 
 # Source Files Quoted if spaced
-SOURCEFILES_QUOTED_IF_SPACED=pic_main.c pingpair_dyn.c serial.c ../../utility/XC8/spi.c ../../RF24_c.c
+SOURCEFILES_QUOTED_IF_SPACED=pic_main.c pingpair_dyn.c serial.c ../../utility/XC8/spi.c ../../RF24.c
 
 # Object Files Quoted if spaced
-OBJECTFILES_QUOTED_IF_SPACED=${OBJECTDIR}/pic_main.p1 ${OBJECTDIR}/pingpair_dyn.p1 ${OBJECTDIR}/serial.p1 ${OBJECTDIR}/_ext/1090826730/spi.p1 ${OBJECTDIR}/_ext/43898991/RF24_c.p1
-POSSIBLE_DEPFILES=${OBJECTDIR}/pic_main.p1.d ${OBJECTDIR}/pingpair_dyn.p1.d ${OBJECTDIR}/serial.p1.d ${OBJECTDIR}/_ext/1090826730/spi.p1.d ${OBJECTDIR}/_ext/43898991/RF24_c.p1.d
+OBJECTFILES_QUOTED_IF_SPACED=${OBJECTDIR}/pic_main.p1 ${OBJECTDIR}/pingpair_dyn.p1 ${OBJECTDIR}/serial.p1 ${OBJECTDIR}/_ext/1090826730/spi.p1 ${OBJECTDIR}/_ext/43898991/RF24.p1
+POSSIBLE_DEPFILES=${OBJECTDIR}/pic_main.p1.d ${OBJECTDIR}/pingpair_dyn.p1.d ${OBJECTDIR}/serial.p1.d ${OBJECTDIR}/_ext/1090826730/spi.p1.d ${OBJECTDIR}/_ext/43898991/RF24.p1.d
 
 # Object Files
-OBJECTFILES=${OBJECTDIR}/pic_main.p1 ${OBJECTDIR}/pingpair_dyn.p1 ${OBJECTDIR}/serial.p1 ${OBJECTDIR}/_ext/1090826730/spi.p1 ${OBJECTDIR}/_ext/43898991/RF24_c.p1
+OBJECTFILES=${OBJECTDIR}/pic_main.p1 ${OBJECTDIR}/pingpair_dyn.p1 ${OBJECTDIR}/serial.p1 ${OBJECTDIR}/_ext/1090826730/spi.p1 ${OBJECTDIR}/_ext/43898991/RF24.p1
 
 # Source Files
-SOURCEFILES=pic_main.c pingpair_dyn.c serial.c ../../utility/XC8/spi.c ../../RF24_c.c
+SOURCEFILES=pic_main.c pingpair_dyn.c serial.c ../../utility/XC8/spi.c ../../RF24.c
 
 
 CFLAGS=
@@ -119,13 +119,13 @@ ${OBJECTDIR}/_ext/1090826730/spi.p1: ../../utility/XC8/spi.c  nbproject/Makefile
 	@-${MV} ${OBJECTDIR}/_ext/1090826730/spi.d ${OBJECTDIR}/_ext/1090826730/spi.p1.d 
 	@${FIXDEPS} ${OBJECTDIR}/_ext/1090826730/spi.p1.d $(SILENT) -rsi ${MP_CC_DIR}../  
 	
-${OBJECTDIR}/_ext/43898991/RF24_c.p1: ../../RF24_c.c  nbproject/Makefile-${CND_CONF}.mk
+${OBJECTDIR}/_ext/43898991/RF24.p1: ../../RF24.c  nbproject/Makefile-${CND_CONF}.mk
 	@${MKDIR} "${OBJECTDIR}/_ext/43898991" 
-	@${RM} ${OBJECTDIR}/_ext/43898991/RF24_c.p1.d 
-	@${RM} ${OBJECTDIR}/_ext/43898991/RF24_c.p1 
-	${MP_CC} --pass1 $(MP_EXTRA_CC_PRE) --chip=$(MP_PROCESSOR_OPTION) -Q -G  -D__DEBUG=1 --debugger=pickit3  --double=24 --float=24 --emi=wordwrite --opt=default,+asm,+asmfile,-speed,+space,-debug --addrqual=ignore --mode=free -P -N255 -I"../../" -I"../../utility/XC8" --warn=-3 --asmlist -DXPRJ_xc8=$(CND_CONF)  --summary=default,-psect,-class,+mem,-hex,-file --output=default,-inhx032 --runtime=default,+clear,+init,-keep,-no_startup,-download,+config,+clib,-plib $(COMPARISON_BUILD)  --output=-mcof,+elf:multilocs --stack=compiled:auto:auto:auto "--errformat=%f:%l: error: (%n) %s" "--warnformat=%f:%l: warning: (%n) %s" "--msgformat=%f:%l: advisory: (%n) %s"    -o${OBJECTDIR}/_ext/43898991/RF24_c.p1  ../../RF24_c.c 
-	@-${MV} ${OBJECTDIR}/_ext/43898991/RF24_c.d ${OBJECTDIR}/_ext/43898991/RF24_c.p1.d 
-	@${FIXDEPS} ${OBJECTDIR}/_ext/43898991/RF24_c.p1.d $(SILENT) -rsi ${MP_CC_DIR}../  
+	@${RM} ${OBJECTDIR}/_ext/43898991/RF24.p1.d 
+	@${RM} ${OBJECTDIR}/_ext/43898991/RF24.p1 
+	${MP_CC} --pass1 $(MP_EXTRA_CC_PRE) --chip=$(MP_PROCESSOR_OPTION) -Q -G  -D__DEBUG=1 --debugger=pickit3  --double=24 --float=24 --emi=wordwrite --opt=default,+asm,+asmfile,-speed,+space,-debug --addrqual=ignore --mode=free -P -N255 -I"../../" -I"../../utility/XC8" --warn=-3 --asmlist -DXPRJ_xc8=$(CND_CONF)  --summary=default,-psect,-class,+mem,-hex,-file --output=default,-inhx032 --runtime=default,+clear,+init,-keep,-no_startup,-download,+config,+clib,-plib $(COMPARISON_BUILD)  --output=-mcof,+elf:multilocs --stack=compiled:auto:auto:auto "--errformat=%f:%l: error: (%n) %s" "--warnformat=%f:%l: warning: (%n) %s" "--msgformat=%f:%l: advisory: (%n) %s"    -o${OBJECTDIR}/_ext/43898991/RF24.p1  ../../RF24.c 
+	@-${MV} ${OBJECTDIR}/_ext/43898991/RF24.d ${OBJECTDIR}/_ext/43898991/RF24.p1.d 
+	@${FIXDEPS} ${OBJECTDIR}/_ext/43898991/RF24.p1.d $(SILENT) -rsi ${MP_CC_DIR}../  
 	
 else
 ${OBJECTDIR}/pic_main.p1: pic_main.c  nbproject/Makefile-${CND_CONF}.mk
@@ -160,13 +160,13 @@ ${OBJECTDIR}/_ext/1090826730/spi.p1: ../../utility/XC8/spi.c  nbproject/Makefile
 	@-${MV} ${OBJECTDIR}/_ext/1090826730/spi.d ${OBJECTDIR}/_ext/1090826730/spi.p1.d 
 	@${FIXDEPS} ${OBJECTDIR}/_ext/1090826730/spi.p1.d $(SILENT) -rsi ${MP_CC_DIR}../  
 	
-${OBJECTDIR}/_ext/43898991/RF24_c.p1: ../../RF24_c.c  nbproject/Makefile-${CND_CONF}.mk
+${OBJECTDIR}/_ext/43898991/RF24.p1: ../../RF24.c  nbproject/Makefile-${CND_CONF}.mk
 	@${MKDIR} "${OBJECTDIR}/_ext/43898991" 
-	@${RM} ${OBJECTDIR}/_ext/43898991/RF24_c.p1.d 
-	@${RM} ${OBJECTDIR}/_ext/43898991/RF24_c.p1 
-	${MP_CC} --pass1 $(MP_EXTRA_CC_PRE) --chip=$(MP_PROCESSOR_OPTION) -Q -G  --double=24 --float=24 --emi=wordwrite --opt=default,+asm,+asmfile,-speed,+space,-debug --addrqual=ignore --mode=free -P -N255 -I"../../" -I"../../utility/XC8" --warn=-3 --asmlist -DXPRJ_xc8=$(CND_CONF)  --summary=default,-psect,-class,+mem,-hex,-file --output=default,-inhx032 --runtime=default,+clear,+init,-keep,-no_startup,-download,+config,+clib,-plib $(COMPARISON_BUILD)  --output=-mcof,+elf:multilocs --stack=compiled:auto:auto:auto "--errformat=%f:%l: error: (%n) %s" "--warnformat=%f:%l: warning: (%n) %s" "--msgformat=%f:%l: advisory: (%n) %s"    -o${OBJECTDIR}/_ext/43898991/RF24_c.p1  ../../RF24_c.c 
-	@-${MV} ${OBJECTDIR}/_ext/43898991/RF24_c.d ${OBJECTDIR}/_ext/43898991/RF24_c.p1.d 
-	@${FIXDEPS} ${OBJECTDIR}/_ext/43898991/RF24_c.p1.d $(SILENT) -rsi ${MP_CC_DIR}../  
+	@${RM} ${OBJECTDIR}/_ext/43898991/RF24.p1.d 
+	@${RM} ${OBJECTDIR}/_ext/43898991/RF24.p1 
+	${MP_CC} --pass1 $(MP_EXTRA_CC_PRE) --chip=$(MP_PROCESSOR_OPTION) -Q -G  --double=24 --float=24 --emi=wordwrite --opt=default,+asm,+asmfile,-speed,+space,-debug --addrqual=ignore --mode=free -P -N255 -I"../../" -I"../../utility/XC8" --warn=-3 --asmlist -DXPRJ_xc8=$(CND_CONF)  --summary=default,-psect,-class,+mem,-hex,-file --output=default,-inhx032 --runtime=default,+clear,+init,-keep,-no_startup,-download,+config,+clib,-plib $(COMPARISON_BUILD)  --output=-mcof,+elf:multilocs --stack=compiled:auto:auto:auto "--errformat=%f:%l: error: (%n) %s" "--warnformat=%f:%l: warning: (%n) %s" "--msgformat=%f:%l: advisory: (%n) %s"    -o${OBJECTDIR}/_ext/43898991/RF24.p1  ../../RF24.c 
+	@-${MV} ${OBJECTDIR}/_ext/43898991/RF24.d ${OBJECTDIR}/_ext/43898991/RF24.p1.d 
+	@${FIXDEPS} ${OBJECTDIR}/_ext/43898991/RF24.p1.d $(SILENT) -rsi ${MP_CC_DIR}../  
 	
 endif
