+++ examples_xc8/pingpair_dyn.X/nbproject/private/private.xml
@@ -2,6 +2,9 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/home/gamboa/Arduino/rf24libs/RF24Ethernet_c/examples_xc8/InteractiveServer_Mesh.X/InteractiveServer_Mesh.c</file>
+            <file>file:/home/gamboa/Arduino/rf24libs/RF24Ethernet_c/RF24Ethernet_c.h</file>
+        </group>
     </open-files>
 </project-private>
