+++ examples_xc8/pingpair_dyn.X/pingpair_dyn.c
@@ -13,7 +13,7 @@
  */
 
 #include "nRF24L01.h"
-#include "RF24_c.h"
+#include "RF24.h"
 
 #include "serial.h"
