+++ library.properties
@@ -1,4 +1,4 @@
-name=RF24_c
+name=RF24
 version=1.2.0
 author=TMRh20
 maintainer=TMRh20
